use int instead of uint, who cares

+++ code.c
@@ -7,13 +7,13 @@
 int main()
 {
     // Variables
-    uint8_t  destinationTileFrom[101];
-    uint8_t  playerPositions[PLAYER_COUNT];
-    uint32_t currentPlayerIndex;
+    int destinationTileFrom[101];
+    int playerPositions[PLAYER_COUNT];
+    int currentPlayerIndex;
 
 
     // init
-    for (uint32_t i=0; i < 101; ++i)
+    for (int i=0; i < 101; ++i)
     {
         destinationTileFrom[i] = i;
     }
@@ -28,7 +28,7 @@ int main()
     ...
 
 
-    for (uint32_t i=0; i < PLAYER_COUNT; ++i)
+    for (int i=0; i < PLAYER_COUNT; ++i)
     {
         playerPositions[i] = 0;
     }
@@ -39,28 +39,28 @@ int main()
     // Game Loop
     while (true)
     {
-        uint32_t diceRoll = random_int(1, 6);
-        printf("Player %u rolled a %u while being on case %u\n", currentPlayerIndex + 1, diceRoll, playerPositions[ currentPlayerIndex ]);
+        int diceRoll = random_int(1, 6);
+        printf("Player %d rolled a %d while being on case %d\n", currentPlayerIndex + 1, diceRoll, playerPositions[ currentPlayerIndex ]);
 
-        uint32_t tileAfterDiceRoll = playerPositions[ currentPlayerIndex ] + diceRoll;
+        int tileAfterDiceRoll = playerPositions[ currentPlayerIndex ] + diceRoll;
         playerPositions[ currentPlayerIndex ] = destinationTileFrom[ tileAfterDiceRoll ];
 
         if (destinationTileFrom[ tileAfterDiceRoll ] > tileAfterDiceRoll)
         {
-            printf("Player %u landed on %u BUT took a ladder to %u\n", currentPlayerIndex + 1, tileAfterDiceRoll, destinationTileFrom[ tileAfterDiceRoll ]);
+            printf("Player %d landed on %d BUT took a ladder to %d\n", currentPlayerIndex + 1, tileAfterDiceRoll, destinationTileFrom[ tileAfterDiceRoll ]);
         }
         else if (destinationTileFrom[ tileAfterDiceRoll ] < tileAfterDiceRoll)
         {
-            printf("Player %u landed on %u BUT took a snake to %u\n", currentPlayerIndex + 1, tileAfterDiceRoll, destinationTileFrom[ tileAfterDiceRoll ]);
+            printf("Player %d landed on %d BUT took a snake to %d\n", currentPlayerIndex + 1, tileAfterDiceRoll, destinationTileFrom[ tileAfterDiceRoll ]);
         }
         else if (destinationTileFrom[ tileAfterDiceRoll ] == 100)
         {
-            printf("Player %u WON !!!!!", currentPlayerIndex + 1);
+            printf("Player %d WON !!!!!", currentPlayerIndex + 1);
             return 0;
         }
         else
         {
-            printf("Player %u landed %u\n", currentPlayerIndex + 1, destinationTileFrom[ tileAfterDiceRoll ]);
+            printf("Player %d landed %u\n", currentPlayerIndex + 1, destinationTileFrom[ tileAfterDiceRoll ]);
         }
         
         currentPlayerIndex = (currentPlayerIndex + 1) % PLAYER_COUNT;
